back/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-sqlalchemy==3.0.5
sqlalchemy==2.0.21
psycopg2-binary==2.9.9
python-dotenv==1.0.0
PyYAML==6.0.1
PyJWT==2.8.0
bcrypt==4.0.1


